+++ lib/components/chat_card.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:kakao/models/user.dart';
 
 import '../models/chat.dart';
+import '../screens/chat_room_screen.dart';
 
 class ChatCard extends StatelessWidget {
   const ChatCard({
@@ -75,177 +74,6 @@ class ChatCard extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class ChatRoomScreen extends StatelessWidget {
-  final chats = [];
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: const Color(0xFFB2C7DA),
-      child: Scaffold(
-        // ✅기본 배경이 흰색이기 때문에 transparent
-        backgroundColor: Colors.transparent,
-        appBar: AppBar(
-          backgroundColor: Colors.transparent,
-          title: Text(
-            "홍길동",
-            style: Theme.of(context).textTheme.headline6,
-          ),
-          actions: const [
-            Icon(FontAwesomeIcons.search, size: 20),
-            SizedBox(width: 20),
-            Icon(FontAwesomeIcons.bars, size: 20),
-            SizedBox(width: 20),
-          ],
-        ),
-        body: Column(
-          children: [
-            Expanded(
-              child: SingleChildScrollView(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                  child: Column(
-                    children: [
-                      TimeLine(time: "2024년 4월 17일 금요일"),
-                      OtherChat(
-                        name: "최주호",
-                        text: "많이 늘었네",
-                        time: "오후 2:50",
-                      ),
-                      MyChat(
-                        text: "ㄟ( ▔, ▔ )ㄏ",
-                        time: "오후 2:51",
-                      ),
-                      ...List.generate(
-                        chats.length,
-                        (index) => chats[index],
-                      )
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            Container(
-              height: 60,
-              color: Colors.white,
-              child: Row(
-                children: [
-                  IconButton(
-                    padding: EdgeInsets.symmetric(horizontal: 15.0),
-                    onPressed: () {},
-                    icon: Icon(FontAwesomeIcons.ad),
-                    iconSize: 25,
-                  ),
-                ],
-              ),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class MyChat extends StatelessWidget {
-  const MyChat({
-    super.key,
-    this.text,
-    this.time,
-  });
-
-  final text;
-  final time;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(10.0),
-      child: Row(
-        // 내 채팅이니까 오른쪽 끝
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: [
-          Text(time, style: const TextStyle(fontSize: 12)),
-          SizedBox(width: 5),
-          Flexible(
-              child: Container(
-            padding: EdgeInsets.all(8.0),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(13.0),
-              color: Color(0xFFFEEC34),
-            ),
-            child: Text(text),
-          ))
-        ],
-      ),
-    );
-  }
-}
-
-class TimeLine extends StatelessWidget {
-  const TimeLine({
-    super.key,
-    required this.time,
-  });
-
-  final String time;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.all(7.0),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(15),
-        color: Color(0xFF9CAFBE),
-      ),
-      child: Text(
-        time,
-        style: TextStyle(color: Colors.white),
-      ),
-    );
-  }
-}
-
-class OtherChat extends StatelessWidget {
-  const OtherChat({
-    super.key,
-    required this.name,
-    required this.text,
-    required this.time,
-  });
-
-  final String name;
-  final String text;
-  final String time;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        CircleAvatar(
-          backgroundImage: NetworkImage(friends[0].backgroundImage),
-        ),
-        SizedBox(
-          width: 10,
-        ),
-        Flexible(
-            child: Column(
-          children: [
-            Text(name),
-            Container(
-              child: Text(text),
-              padding: EdgeInsets.all(8),
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(13),
-                color: Colors.white,
-              ),
-            )
-          ],
-        ))
-      ],
     );
   }
 }
